bms_base/backend/endpoint/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.6.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
concurrent-log-handler==0.9.20
cryptography==35.0.0
Deprecated==1.2.13
Flask==1.1.4
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.2
idna==3.3
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.1.2
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
passlib==1.7.4
pendulum==2.1.2
portalocker==2.5.1
psycopg2==2.9.1
py-buzz==1.0.3
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.2.0
pyOpenSSL==21.0.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytzdata==2020.1
pywin32==304
requests==2.26.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
urllib3==1.26.7
Werkzeug==1.0.1
wrapt==1.13.2
